% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{download_assessment_historical}
\alias{download_assessment_historical}
\title{Download Historical Assessment Data (2012-2019)}
\usage{
download_assessment_historical(end_year)
}
\arguments{
\item{end_year}{School year end (2012-2019)}
}
\value{
List with assessment data frames
}
\description{
Downloads assessment data from SRC Historical Datasets.
}
\keyword{internal}
