% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{download_kde_assessment_csv}
\alias{download_kde_assessment_csv}
\title{Download a CSV assessment file from KDE}
\usage{
download_kde_assessment_csv(url, end_year, file_type)
}
\arguments{
\item{url}{URL to download}

\item{end_year}{School year for error messages}

\item{file_type}{File type for error messages}
}
\value{
Data frame
}
\description{
Download a CSV assessment file from KDE
}
\keyword{internal}
