% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_saar_data}
\alias{download_saar_data}
\title{Download SAAR enrollment data (1997-2019)}
\usage{
download_saar_data(end_year)
}
\arguments{
\item{end_year}{School year end (1997-2019)}
}
\value{
List with district enrollment data frame
}
\description{
Downloads SAAR Ethnic Membership Report data from KDE.
This is district-level only.
}
\keyword{internal}
