% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_saar_single_year}
\alias{download_saar_single_year}
\title{Download individual year SAAR report}
\usage{
download_saar_single_year(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with SAAR data
}
\description{
Downloads a single-year SAAR report for more recent years.
}
\keyword{internal}
