% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_src_historical}
\alias{download_src_historical}
\title{Download SRC Historical data (2012-2019)}
\usage{
download_src_historical(end_year)
}
\arguments{
\item{end_year}{School year end (2012-2019)}
}
\value{
List with enrollment data frames
}
\description{
Downloads enrollment data from SRC Historical Datasets.
}
\keyword{internal}
