% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{get_assessment_urls}
\alias{get_assessment_urls}
\title{Get the URL for KDE assessment data}
\usage{
get_assessment_urls(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Character vector of URLs
}
\description{
Returns the URL(s) where assessment data can be found for a given year.
}
\examples{
get_assessment_urls(2024)
get_assessment_urls(2015)
}
