% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_enrollment_urls}
\alias{get_enrollment_urls}
\title{Get the URL for KDE enrollment data}
\usage{
get_enrollment_urls(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Character vector of URLs
}
\description{
Returns the URL(s) where enrollment data can be found for a given year.
}
\examples{
get_enrollment_urls(2024)  # KYRC24 format
get_enrollment_urls(2023)  # primary/secondary format
get_enrollment_urls(2005)  # SAAR format
}
