% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{get_raw_assessment}
\alias{get_raw_assessment}
\title{Download raw assessment data from KDE}
\usage{
get_raw_assessment(end_year)
}
\arguments{
\item{end_year}{School year end (2012-2025)}
}
\value{
List with assessment data frames
}
\description{
Downloads assessment data from Kentucky Department of Education.
Uses SRC datasets for all available years (2011-2012 to present).
}
\keyword{internal}
