% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_assessment.R
\name{id_assessment_aggs}
\alias{id_assessment_aggs}
\title{Create district and school aggregates for tidy assessment data}
\usage{
id_assessment_aggs(df)
}
\arguments{
\item{df}{Tidy assessment data frame}
}
\value{
Data frame with added district aggregates
}
\description{
Aggregates school-level data to district level.
}
\keyword{internal}
