% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{process_assessment}
\alias{process_assessment}
\title{Process raw KDE assessment data}
\usage{
process_assessment(raw_data, end_year)
}
\arguments{
\item{raw_data}{List containing data frames from get_raw_assessment}

\item{end_year}{School year end}
}
\value{
Processed data frame with standardized columns
}
\description{
Transforms raw assessment data into a standardized schema.
}
\keyword{internal}
