% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{process_assessment_df}
\alias{process_assessment_df}
\title{Process a single SRC assessment data frame}
\usage{
process_assessment_df(df, end_year)
}
\arguments{
\item{df}{Raw data frame from SRC assessment data}

\item{end_year}{School year end}
}
\value{
Processed data frame in wide format
}
\description{
Process a single SRC assessment data frame
}
\keyword{internal}
