% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_saar}
\alias{process_saar}
\title{Process SAAR data (1997-2019)}
\usage{
process_saar(raw_data, end_year)
}
\arguments{
\item{raw_data}{List with district data frame}

\item{end_year}{School year end}
}
\value{
Processed data frame
}
\description{
Process SAAR data (1997-2019)
}
\keyword{internal}
