% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_src_current}
\alias{process_src_current}
\title{Process SRC Current Format data (2020+)}
\usage{
process_src_current(raw_data, end_year)
}
\arguments{
\item{raw_data}{List with primary and secondary data frames (or combined)}

\item{end_year}{School year end}
}
\value{
Processed data frame
}
\description{
Process SRC Current Format data (2020+)
}
\keyword{internal}
