% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_src_historical}
\alias{process_src_historical}
\title{Process SRC Historical data (2012-2019)}
\usage{
process_src_historical(raw_data, end_year)
}
\arguments{
\item{raw_data}{List with enrollment data frames}

\item{end_year}{School year end}
}
\value{
Processed data frame
}
\description{
Process SRC Historical data (2012-2019)
}
\keyword{internal}
