% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{saar_year_to_end_year}
\alias{saar_year_to_end_year}
\title{Get school year end from SAAR year format}
\usage{
saar_year_to_end_year(saar_year)
}
\arguments{
\item{saar_year}{SAAR year value}
}
\value{
School year end (integer)
}
\description{
SAAR uses format like "1999" to mean 1998-99 school year (end_year = 1999).
}
\keyword{internal}
