% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{standardize_assessment_columns}
\alias{standardize_assessment_columns}
\title{Standardize assessment data column names and order}
\usage{
standardize_assessment_columns(df)
}
\arguments{
\item{df}{Data frame to standardize}
}
\value{
Data frame with standardized columns
}
\description{
Standardize assessment data column names and order
}
\keyword{internal}
