% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_assessment.R
\name{tidy_assessment}
\alias{tidy_assessment}
\title{Tidy assessment data}
\usage{
tidy_assessment(df)
}
\arguments{
\item{df}{Processed assessment data frame in wide format}
}
\value{
Data frame in tidy long format
}
\description{
Converts wide-format assessment data into tidy long format with
}
\keyword{internal}
